nie_tracking/visual_tracking/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.9.0=pypi_0
appdirs=1.4.4=pypi_0
blas=1.0=mkl
ca-certificates=2020.1.1=0
cachetools=4.1.0=pypi_0
certifi=2020.4.5.1=py38_0
cffi=1.14.0=pypi_0
chardet=3.0.4=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py38_0
cython=0.29.17=py38he6710b0_0
dbus=1.13.14=hb2f20db_0
decorator=4.4.2=pypi_0
expat=2.2.6=he6710b0_0
filelock=3.0.12=pypi_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
gdown=3.11.0=pypi_0
glib=2.63.1=h5a9c865_0
google-auth=1.14.2=pypi_0
google-auth-oauthlib=0.4.1=pypi_0
grpcio=1.28.1=pypi_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
icu=58.2=he6710b0_3
idna=2.9=pypi_0
imagecodecs=2020.2.18=pypi_0
imageio=2.8.0=pypi_0
intel-openmp=2020.0=166
ipython-genutils=0.2.0=pypi_0
jpeg=9b=h024ee3a_2
jpeg4py=0.1.4=pypi_0
jsonpatch=1.25=pypi_0
jsonpointer=2.0=pypi_0
kiwisolver=1.2.0=py38hfd86e86_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
markdown=3.2.2=pypi_0
matplotlib=3.1.3=py38_0
matplotlib-base=3.1.3=py38hef1b27d_0
mkl=2020.0=166
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.0.15=py38ha843d7b_0
mkl_random=1.1.0=py38h962f231_0
ncurses=6.2=he6710b0_1
networkx=2.4=pypi_0
ninja=1.9.0=py38hfd86e86_0
numpy=1.18.1=py38h4f9e942_0
numpy-base=1.18.1=py38hde5b4d6_1
oauthlib=3.1.0=pypi_0
olefile=0.46=py_0
opencv-python=4.2.0.34=pypi_0
openssl=1.1.1g=h7b6447c_0
packaging=20.3=pypi_0
pandas=1.0.3=py38h0573a6f_0
patsy=0.5.1=pypi_0
pcre=8.43=he6710b0_0
pillow=7.1.2=py38hb39fc2d_0
pip=20.0.2=py38_3
pooch=1.1.0=pypi_0
protobuf=3.11.3=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocotools=2.0.0=pypi_0
pycparser=2.20=pypi_0
pynvml=8.0.4=pypi_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py38h05f1152_4
pysocks=1.7.1=pypi_0
python=3.8.2=hcf32534_0
python-dateutil=2.8.1=py_0
python-utils=2.4.0=pypi_0
pytorch=1.6.0.dev20200512=py3.8_cuda10.2.89_cudnn7.6.5_0
pytz=2020.1=py_0
pywavelets=1.1.1=pypi_0
pyzmq=19.0.1=pypi_0
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
requests=2.23.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.0=pypi_0
scikit-image=0.17.1=pypi_0
scipy=1.4.1=pypi_0
setuptools=46.1.3=py38_0
sip=4.19.13=py38he6710b0_0
six=1.14.0=py38_0
sqlite=3.31.1=h62c20be_1
statsmodels=0.11.1=pypi_0
tb-nightly=2.3.0a20200511=pypi_0
tensorboard-plugin-wit=1.6.0.post3=pypi_0
tensorboardx=2.0=pypi_0
tifffile=2020.5.7=pypi_0
tikzplotlib=0.9.1=pypi_0
tk=8.6.8=hbc83047_0
torchfile=0.1.0=pypi_0
torchvision=0.7.0.dev20200512=py38_cu102
tornado=6.0.4=py38h7b6447c_1
tqdm=4.46.0=py_0
urllib3=1.25.9=pypi_0
utils=1.0.1=pypi_0
visdom=0.1.8.9=pypi_0
websocket-client=0.57.0=pypi_0
werkzeug=1.0.1=pypi_0
wheel=0.34.2=py38_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
